Fix for failing test
